new settings refactoring
